Move container state to entity system Move client code to client
