+++ Robust.Shared/Containers/SharedContainerSystem.cs
@@ -6,8 +6,6 @@ namespace Robust.Shared.Containers
 {
     public abstract class SharedContainerSystem : EntitySystem
     {
-        public readonly Dictionary<EntityUid, IContainer> ExpectedEntities = new();
-
         /// <inheritdoc />
         public override void Initialize()
         {
@@ -26,23 +24,6 @@ namespace Robust.Shared.Containers
 
             if (oldParentEntity.TryGetComponent(out IContainerManager? containerManager))
                 containerManager.ForceRemove(message.Entity);
-        }
-
-        public void AddExpectedEntity(EntityUid uid, IContainer container)
-        {
-            if (ExpectedEntities.ContainsKey(uid))
-                return;
-            ExpectedEntities.Add(uid, container);
-            container.ExpectedEntities.Add(uid);
-        }
-
-        public void RemoveExpectedEntity(EntityUid uid)
-        {
-            if (!ExpectedEntities.ContainsKey(uid))
-                return;
-            var container = ExpectedEntities[uid];
-            ExpectedEntities.Remove(uid);
-            container.ExpectedEntities.Add(uid);
         }
     }
 }
